* Moved cache folder to root folder and added conf option. * Added GuildManager. * Minor initial guild cleaning. * Moved WorldManager.SharpMind to AIScript.SharpMind...

+++ src/WorldServer/Network/Sending/Send.Entity.cs
@@ -481,7 +481,7 @@ namespace Aura.World.Network
 			{
 				packet.PutLong(creature.Guild.Id);
 				packet.PutString(creature.Guild.Name);
-				packet.PutInt(creature.GuildMemberInfo.MemberRank);
+				packet.PutInt((uint)creature.GuildMemberInfo.MemberRank);
 				packet.PutByte(0);
 				packet.PutByte(0);
 				packet.PutByte(0);
